# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Analog API",
    "version": "0.8.0",
    "description": "Shared canvas for a human and their agents. FROZEN CONTRACT — amendments go through WP0 only. Canvas mutations require actor and actor_kind and emit events. The operator upgrade operation changes only the server executable and emits no canvas event. A server started with --ui-only holds no data: it serves /health and the web bundle, and every other operation answers 503 with error ui_only (the UIOnly response)."
  },
  "servers": [{ "url": "http://127.0.0.1:8787/api" }],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http", "scheme": "bearer",
        "description": "A per-actor token. It identifies exactly one actor, and the server rejects an `actor`/`actor_kind` that disagrees with it — attribution in the event log is only worth having if it cannot be claimed. Absent when no tokens are configured, which is the loopback default. Deliberately not a cookie: a card's sandboxed iframe cannot set an Authorization header, so agent-authored HTML cannot ride an ambient credential."
      }
    },
...
    "schemas": {
      "UpgradeStatus": {
        "type": "object",
        "required": ["current", "available", "supported"],
        "properties": {
          "current": { "type": "string" },
          "latest": { "type": "string" },
          "available": { "type": "boolean" },
          "supported": { "type": "boolean" },
          "reason": { "type": "string", "description": "Why self-upgrade is unsupported, or why the latest release could not be checked. A failed lookup returns 200 with supported=true, available=false and this field set." }
        }
      },
...
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Tokens are configured and the request had no valid one.",
        "headers": { "WWW-Authenticate": { "schema": { "type": "string" } } },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
...
    }
  },
  "security": [{ "bearerAuth": [] }, {}],
  "paths": {
...
    "/whoami": {
      "get": {
        "operationId": "whoami",
        "description": "The actor this token writes as. Clients use it instead of assuming a name: the web UI can no longer hardcode 'human' once a token decides who you are.",
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": {
            "type": "object",
            "required": ["authenticated", "actor", "actor_kind"],
            "properties": {
              "authenticated": { "type": "boolean" },
              "actor": { "type": ["string", "null"] },
              "actor_kind": { "type": ["string", "null"], "enum": ["human", "agent", null] }
            }
          } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/upgrade": {
      "get": {
        "operationId": "getUpgrade",
        "description": "Check whether a newer standalone analog-server release is available. Package-manager and service-managed installs report why self-upgrade is unavailable. A release-host lookup failure returns 200 with a reason so an offline home page remains usable.",
        "responses": {
          "200": { "description": "Upgrade status", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UpgradeStatus" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "post": {
        "operationId": "upgradeServer",
        "description": "Download the latest release, verify SHA256SUMS, atomically replace a standalone server binary, and restart the process. Requires a human bearer token when authentication is enabled. Only the server executable changes; canvas data is untouched. A custom header and JSON content type prevent cross-site form submissions on tokenless loopback servers.",
        "parameters": [
          { "name": "X-Analog-Upgrade", "in": "header", "required": true, "schema": { "const": "1" } }
        ],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "maxProperties": 0 } } } },
        "responses": {
          "200": { "description": "Already current", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UpgradeStatus" } } } },
          "202": { "description": "Binary replaced; server is restarting", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UpgradeStatus" } } } },
          "400": { "$ref": "#/components/responses/ValidationFailed" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "500": { "$ref": "#/components/responses/UpgradeFailed" },
          "502": { "$ref": "#/components/responses/UpgradeFailed" }
        }
      }
    },
    "/spaces": {
      "get": {
        "operationId": "listSpaces",
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": {
          "type": "array", "items": { "$ref": "#/components/schemas/Space" } } } } } }
      },
      "post": {
        "operationId": "createSpace",
        "parameters": [
          { "$ref": "#/components/parameters/actor" },
          { "$ref": "#/components/parameters/actorKind" }
        ],
        "requestBody": { "required": true, "content": { "application/json": { "schema": {
          "type": "object", "required": ["slug", "title"],
          "properties": {
            "slug": { "type": "string", "pattern": "^[a-z0-9-]{1,64}$" },
            "title": { "type": "string" },
            "revision_mode": { "type": "string", "enum": ["replace", "branch"], "default": "replace" }
          }
        } } } },
        "responses": {
          "201": { "description": "Created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Space" } } } },
          "400": { "$ref": "#/components/responses/ActorRequired" },
          "409": { "description": "Slug already exists.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/spaces/{slug}": {
      "parameters": [{ "$ref": "#/components/parameters/slug" }],
      "get": {
        "operationId": "getSpace",
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Space" } } } },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "patch": {
        "operationId": "updateSpace",
        "parameters": [
          { "$ref": "#/components/parameters/actor" },
          { "$ref": "#/components/parameters/actorKind" }
        ],
        "requestBody": { "required": true, "content": { "application/json": { "schema": {
          "type": "object",
          "properties": {
            "title": { "type": "string" },
            "revision_mode": { "type": "string", "enum": ["replace", "branch"] }
          }
        } } } },
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Space" } } } },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "operationId": "deleteSpace",
        "parameters": [
          { "$ref": "#/components/parameters/actor" },
          { "$ref": "#/components/parameters/actorKind" }
        ],
        "responses": { "204": { "description": "Deleted" }, "404": { "$ref": "#/components/responses/NotFound" } }
      }
    },
    "/spaces/{slug}/canvas": {
      "parameters": [{ "$ref": "#/components/parameters/slug" }],
      "get": {
        "operationId": "getCanvas",
        "description": "Live cards and links. Soft-deleted rows are excluded. Valid JSON Canvas 1.0.",
        "parameters": [{
          "name": "include_deleted", "in": "query", "required": false,
          "schema": { "type": "boolean", "default": false }
        }],
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Canvas" } } } },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/spaces/{slug}/import": {
      "parameters": [{ "$ref": "#/components/parameters/slug" }],
      "post": {
        "operationId": "importCanvas",
        "description": "Additive only. Creates cards and links; never updates or deletes. Incoming ids are remapped to fresh ones. Emits one card.created / link.created event per item.",
        "parameters": [
          { "$ref": "#/components/parameters/actor" },
          { "$ref": "#/components/parameters/actorKind" }
        ],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Canvas" } } } },
        "responses": {
          "201": { "description": "Imported", "content": { "application/json": { "schema": {
            "type": "object",
            "properties": {
              "id_map": { "type": "object", "additionalProperties": { "type": "string" } },
              "canvas": { "$ref": "#/components/schemas/Canvas" }
            }
          } } } },
          "400": { "$ref": "#/components/responses/ActorRequired" }
        }
      }
    },
    "/spaces/{slug}/cards": {
      "parameters": [{ "$ref": "#/components/parameters/slug" }],
      "post": {
        "operationId": "createCards",
        "description": "Bulk create. Accepts either agent-friendly drafts or raw JSON Canvas nodes.",
        "parameters": [
          { "$ref": "#/components/parameters/actor" },
          { "$ref": "#/components/parameters/actorKind" }
        ],
        "requestBody": { "required": true, "content": { "application/json": { "schema": {
          "type": "object",
          "oneOf": [
            { "required": ["cards"], "properties": { "cards": { "type": "array", "items": { "$ref": "#/components/schemas/CardDraft" } } } },
            { "required": ["nodes"], "properties": { "nodes": { "type": "array", "items": { "$ref": "#/components/schemas/Node" } } } }
          ]
        } } } },
        "responses": {
          "201": { "description": "Created", "content": { "application/json": { "schema": {
            "type": "array", "items": { "$ref": "#/components/schemas/Node" } } } } },
          "400": { "$ref": "#/components/responses/ActorRequired" }
        }
      }
    },
    "/spaces/{slug}/cards/{card_id}": {
      "parameters": [
        { "$ref": "#/components/parameters/slug" },
        { "name": "card_id", "in": "path", "required": true, "schema": { "type": "string" } }
      ],
      "patch": {
        "operationId": "updateCard",
        "description": "Partial node update. Geometry-only changes emit card.moved and do NOT bump sp_rev. Any other change emits card.updated and bumps sp_rev. In branch mode a content change instead emits TWO events — card.created for the replacement and link.created for the auto-link labelled 'revised' — and no card.updated, so the superseded card's sp_rev stays frozen and its annotations can never go stale.",
        "parameters": [
          { "$ref": "#/components/parameters/actor" },
          { "$ref": "#/components/parameters/actorKind" },
          {
            "name": "mode", "in": "query", "required": false,
            "description": "Overrides the space's revision_mode for this call.",
            "schema": { "type": "string", "enum": ["replace", "branch"] }
          },
          {
            "name": "If-Match", "in": "header", "required": false,
            "description": "Expected sp_rev. On mismatch returns 409 with the current node.",
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": { "required": true, "content": { "application/json": { "schema": {
          "type": "object", "additionalProperties": true,
          "description": "Any subset of Node properties."
        } } } },
        "responses": {
          "200": { "description": "Updated. In branch mode this is the NEW card; the superseded one carries sp_superseded_by.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Node" } } } },
          "400": { "$ref": "#/components/responses/ActorRequired" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "description": "If-Match did not match current sp_rev.", "content": { "application/json": { "schema": {
            "type": "object",
            "properties": {
              "error": { "const": "conflict" },
              "message": { "type": "string" },
              "current": { "$ref": "#/components/schemas/Node" }
            }
          } } } }
        }
      },
      "delete": {
        "operationId": "deleteCard",
        "description": "Soft delete. The row is retained and appears in feedback as cards_deleted.",
        "parameters": [
          { "$ref": "#/components/parameters/actor" },
          { "$ref": "#/components/parameters/actorKind" }
        ],
        "responses": { "204": { "description": "Deleted" }, "404": { "$ref": "#/components/responses/NotFound" } }
      }
    },
    "/spaces/{slug}/links": {
      "parameters": [{ "$ref": "#/components/parameters/slug" }],
      "post": {
        "operationId": "createLinks",
        "parameters": [
          { "$ref": "#/components/parameters/actor" },
          { "$ref": "#/components/parameters/actorKind" }
        ],
        "requestBody": { "required": true, "content": { "application/json": { "schema": {
          "type": "object", "required": ["edges"],
          "properties": { "edges": { "type": "array", "items": {
            "type": "object", "required": ["fromNode", "toNode"],
            "properties": {
              "fromNode": { "type": "string" }, "toNode": { "type": "string" },
              "fromSide": { "type": "string" }, "toSide": { "type": "string" },
              "label": { "type": "string" }, "color": { "type": "string" }
            }
          } } }
        } } } },
        "responses": {
          "201": { "description": "Created", "content": { "application/json": { "schema": {
            "type": "array", "items": { "$ref": "#/components/schemas/Edge" } } } } },
          "400": { "$ref": "#/components/responses/ActorRequired" },
          "404": { "description": "fromNode or toNode does not exist.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/spaces/{slug}/links/{link_id}": {
      "parameters": [
        { "$ref": "#/components/parameters/slug" },
        { "name": "link_id", "in": "path", "required": true, "schema": { "type": "string" } }
      ],
      "delete": {
        "operationId": "deleteLink",
        "parameters": [
          { "$ref": "#/components/parameters/actor" },
          { "$ref": "#/components/parameters/actorKind" }
        ],
        "responses": { "204": { "description": "Deleted" }, "404": { "$ref": "#/components/responses/NotFound" } }
      }
    },
    "/spaces/{slug}/annotations": {
      "parameters": [{ "$ref": "#/components/parameters/slug" }],
      "get": {
        "operationId": "listAnnotations",
        "parameters": [
          { "name": "resolved", "in": "query", "schema": { "type": "boolean" } },
          { "name": "card_id", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": {
          "type": "array", "items": { "$ref": "#/components/schemas/Annotation" } } } } } }
      },
      "post": {
        "operationId": "createAnnotation",
        "parameters": [
          { "$ref": "#/components/parameters/actor" },
          { "$ref": "#/components/parameters/actorKind" }
        ],
        "requestBody": { "required": true, "content": { "application/json": { "schema": {
          "type": "object", "required": ["card_id", "body"],
          "properties": {
            "card_id": { "type": "string" },
            "selector": { "$ref": "#/components/schemas/Selector" },
            "body": { "type": "string" },
            "motivation": { "type": "string", "enum": ["commenting", "assessing", "editing"], "default": "commenting" }
          }
        } } } },
        "responses": {
          "201": { "description": "Created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Annotation" } } } },
          "400": { "$ref": "#/components/responses/ActorRequired" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/spaces/{slug}/annotations/{annotation_id}": {
      "parameters": [
        { "$ref": "#/components/parameters/slug" },
        { "name": "annotation_id", "in": "path", "required": true, "schema": { "type": "string" } }
      ],
      "patch": {
        "operationId": "resolveAnnotation",
        "parameters": [
          { "$ref": "#/components/parameters/actor" },
          { "$ref": "#/components/parameters/actorKind" }
        ],
        "requestBody": { "required": true, "content": { "application/json": { "schema": {
          "type": "object",
          "properties": { "resolved": { "type": "boolean" }, "reply": { "type": "string" } }
        } } } },
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Annotation" } } } },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
  }
}